Refactor count modifier move it to the "features" folder
